feat: Introduce Transaction struct and related methods
- Added a new Transaction struct in `transaction.rs` with fields for transaction ID, sender, receiver, amount, gas fee, timestamp, signature, and optional data.
- Implemented methods for signing, verifying, and categorizing transactions.
- Updated the VM module to utilize the new Transaction struct for blockchain interactions.
- Removed the tps_benchmark.rs file as it was no longer needed.
- Refactored various modules to replace direct blockchain transaction references with the new Transaction struct.
- Enhanced logging and error handling throughout the transaction processing flow.
