+++ crates/command/src/keytool_cli/mod.rs
@@ -7,7 +7,7 @@ use mona_crypto::{
 };
 use std::io::{self, Write};
 
-use mona_blockchain::blockchain::{get_balance, load_blockchain_with_retry};
+use mona_types::storage::{get_balance, load_blockchain_with_retry};
 use rpassword::read_password;
 use std::process::exit;
 use serde_json::json;
