+++ crates/rpc/rpc-api/src/accounts.rs
@@ -1,9 +1,10 @@
 use std::str::FromStr;
 use jsonrpc_core::{Params, Result as JsonRpcResult, Error as RpcError, ErrorCode};
 
+use mona_blockchain::CHAIN_ID;
 use mona_crypto::{list_wallet_files, load_wallet};
 use mona_types::address::Address;
-use mona_blockchain::{blockchain::{load_blockchain_with_retry, BALANCES, BLOCKCHAIN_DATA}, chain_id::CHAIN_ID};
+use mona_types::storage::{load_blockchain_with_retry, BALANCES, BLOCKCHAIN_DATA};
 use panorama::simulation::process_transfer;
 use panorama::utils::format_kari_amount;
 use serde_json::{json, Value as JsonValue};
