+++ monaos/mona-blockchain/Cargo.toml
@@ -12,47 +12,3 @@ repository.workspace = true
 description.workspace = true
 
 [dependencies]
-lazy_static.workspace = true
-once_cell.workspace = true  # Add once_cell dependency
-clap.workspace = true  # Add clap for CLI argument parsing
-
-# Cryptography
-sha2.workspace = true
-rand.workspace = true
-bip39.workspace = true
-sha3.workspace = true
-blake3 = "1.3.1"  # Add direct blake3 dependency for clarity
-
-# Serialization
-serde.workspace = true
-serde_json.workspace = true
-bincode.workspace = true
-bcs.workspace = true
-
-# Database
-rocksdb.workspace = true
-
-# Networking
-jsonrpc-core.workspace = true
-jsonrpc-http-server.workspace = true
-futures.workspace = true
-tokio.workspace = true
-serde_yaml.workspace = true
-
-# Utilities
-digest.workspace = true
-hex.workspace = true
-chrono.workspace = true
-dirs.workspace = true
-log.workspace = true    
-env_logger.workspace = true
-colored.workspace = true
-thiserror.workspace = true
-
-# Internal dependencies
-key.workspace = true
-mona-crypto = { workspace = true, features = ["blake3"] }  # Add blake3 feature
-mona-types.workspace = true
-common.workspace = true
-consensus-pos.workspace = true
-mona-storage.workspace = true
