+++ monaos/mona-blockchain/src/lib.rs
@@ -1,4 +1,7 @@
-pub mod blockchain;
-pub mod block;
 pub mod chain_id;
+
+
+// pub use address::Address; // Commented out until Address is properly defined
+pub use chain_id::CHAIN_ID;
+
