+++ monaos/mona-types/Cargo.toml
@@ -24,6 +24,8 @@ move-vm-types = { workspace = true, optional = true }
 move-vm-profiler = { workspace = true, optional = true }
 
 
+serde_json = { workspace = true }
+dirs.workspace = true
 mona-enum-compat-util.workspace = true
 log.workspace = true
 lazy_static.workspace = true
@@ -34,6 +36,10 @@ sha3.workspace = true
 hex.workspace = true
 primitive-types.workspace = true
 thiserror.workspace = true
+bincode.workspace = true
+mona-storage.workspace = true
+consensus-pos.workspace = true
+mona-blockchain.workspace = true
 
 [features]
 # This feature is used to include the gas profiler in the mona-types crate.
